Creating front and backend
Backend:
Did the connection with the database
Created the first controllers in the backend for signing up and in.
Created the model for the users collection and the routes for theses controllers.

Frontend:
Created the LandPage with its header, first buttons and styling
Created the router.js and created this first route